feat: ✨ move annotator under src folder and create better folder structure
